requirements.txt
pandas
tensorflow
numpy
matplotlib
sklearn
pyyaml
termcolor
itertools
